requirement.txt
crew
langchain_google_gemini
load_dotenv
typing
python-telegram-bot
fpdf
